+++ .github/workflows/build.yml
@@ -1,6 +1,7 @@
 name: "Build and Report Generation"
 on: 
   push:
+    branches : [main]
     paths:
       - '**'
       - '!docs/**'
@@ -22,11 +23,31 @@ jobs:
         with:
           submodules: recursive
       - name: Use Java ${{ matrix.java }}
-        uses: actions/setup-java@v1
+        uses: actions/setup-java@v2
         with:
+          distribution: 'zulu'
           java-version: ${{ matrix.java }}
+          cache: gradle
       - name: Build and test
-        run: ./gradlew build
+        run: ./gradlew build -x fuzzTest
+
+  fuzz-test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        java: [ 11, 17 ]
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          submodules: recursive
+      - name: Use Java ${{ matrix.java }}
+        uses: actions/setup-java@v2
+        with:
+          distribution: 'zulu'
+          java-version: ${{ matrix.java }}
+          cache: gradle
+      - name: Build and test
+        run: ./gradlew fuzzTest
 
   report-generation:
     runs-on: ubuntu-latest
@@ -35,9 +56,11 @@ jobs:
         with:
           submodules: recursive
       - name: Use Java 17
-        uses: actions/setup-java@v1
+        uses: actions/setup-java@v2
         with:
+          distribution: 'zulu'
           java-version: 17
+          cache: gradle
       - name: Generate test coverage report
         run: ./gradlew build jacocoTestReport
